* Moving CalibComponents into calib_components.py. * Improved tests * Fixed a few bugs
